pyproject.toml
[tool.poetry]
name = "cinemaauth"
version = "0.1.0"
description = "Сервис авторизации пользователей кинотеатра"
authors = ["Yandex Practicum Team"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
uvicorn = "0.23.2"
gunicorn = "21.2.0"
fastapi = "0.100.1"
pydantic = "2.4.2"
pydantic-settings = "2.0.3"
aioredis = "2.0.1"
orjson = "3.9.9"
SQLAlchemy = "2.0.23"
alembic = "1.12.1"
async-fastapi-jwt-auth = '0.6.1'
typer = "0.9.0"
passlib = "1.7.4"
python-jose = "^3.3.0"

[tool.poetry.dev-dependencies]
Faker = "14.0.0"
pytest = "7.0.1"
mypy = "1.1.1"
ruff = "0.1.6"

[tool.ruff]
# Exclude a variety of commonly ignored directories.
exclude = [
    ".bzr",
    ".direnv",
    ".eggs",
    ".git",
    ".git-rewrite",
    ".hg",
    ".mypy_cache",
    ".nox",
    ".pants.d",
    ".pytype",
    ".ruff_cache",
    ".svn",
    ".tox",
    ".venv",
    "__pypackages__",
    "_build",
    "buck-out",
    "build",
    "dist",
    "node_modules",
    "venv",
]

# Same as Black.
line-length = 88
indent-width = 4

# Assume Python 3.8
target-version = "py311"

[tool.ruff.lint]
# Enable Pyflakes (`F`) and a subset of the pycodestyle (`E`)  codes by default.
select = ["E4", "E7", "E9", "F"]
ignore = []

# Allow fix for all enabled rules (when `--fix`) is provided.
fixable = ["ALL"]
unfixable = []

# Allow unused variables when underscore-prefixed.
dummy-variable-rgx = "^(_+|(_+[a-zA-Z0-9_]*[a-zA-Z0-9]+?))$"

[tool.ruff.format]
# Like Black, use double quotes for strings.
quote-style = "double"

# Like Black, indent with spaces, rather than tabs.
indent-style = "space"

# Like Black, respect magic trailing commas.
skip-magic-trailing-comma = false

# Like Black, automatically detect the appropriate line ending.
line-ending = "auto"

[tool.mypy]
plugins = ["pydantic.mypy"]
follow_imports = "silent"
warn_redundant_casts = true
warn_unused_ignores = true
disallow_any_generics = true
check_untyped_defs = true
no_implicit_reexport = true
disallow_untyped_defs = true

[tool.pydantic-mypy]
init_forbid_extra = true
init_typed = true
warn_required_dynamic_aliases = true